% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_reaction_table.R
\name{collapse_reactions}
\alias{collapse_reactions}
\title{parse a long format metabolic model to a gurobi model}
\usage{
collapse_reactions(reactions_expanded, reaction_table)
}
\arguments{
\item{reactions_expanded}{A data frame as output by \code{expand_reactions}}

\item{reaction_table}{A data frame describing the metabolic model.}
}
\description{
Used as the second half of \code{parse_reactions}, this parses the long format produced by \code{expand_reactions} to a gurobi model
}
\details{
The \code{reaction_table} must have columns:
\itemize{
 \item \code{abbreviation},
 \item \code{equation},
 \item \code{uppbnd},
 \item \code{lowbnd}, and
 \item \code{obj_coef}.
}
}

