% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_reaction_table.R
\name{expand_reactions}
\alias{expand_reactions}
\title{parse reaction table to an intermediate, long format}
\usage{
expand_reactions(reaction_table, pattern_arrow = "<?[-=]+>")
}
\arguments{
\item{reaction_table}{A data frame describing the metabolic model.}
}
\description{
Used as the first half of \code{parse_reactions}
}
\details{
The \code{reaction_table} must have columns:
\itemize{
 \item \code{abbreviation},
 \item \code{equation},
 \item \code{uppbnd},
 \item \code{lowbnd}, and
 \item \code{obj_coef}.
}
}

