% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsing_and_conversion.R
\name{expanded_to_reactiontbl}
\alias{expanded_to_reactiontbl}
\title{Convert intermediate expanded format back to a reaction table}
\usage{
expanded_to_reactiontbl(expanded)
}
\arguments{
\item{expanded}{A list of data frames: \itemize{
  \item \code{rxns}, which has one row per reaction, 
  \item \code{mets}, which has one row for each metabolite, and 
  \item \code{stoich}, which has one row for each time a metabolite appears in a reaction.
}}
}
\value{
A data frame describing the metabolic model.
}
\description{
Useful for saving a new or edited model
}

