% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsing_and_conversion.R
\name{expanded_to_ROI}
\alias{expanded_to_ROI}
\title{Parse a long format metabolic model to an ROI model}
\usage{
expanded_to_ROI(reactions_expanded)
}
\arguments{
\item{reactions_expanded}{A list of data frames as output by \code{reactiontbl_to_expanded}}
}
\value{
A list suitable for input to ROI.
}
\description{
This parses the long format produced by \code{reactiontbl_to_expanded} to an ROI model
}
\details{
The \code{reaction_table} must have columns:
\itemize{
 \item \code{abbreviation},
 \item \code{equation},
 \item \code{uppbnd},
 \item \code{lowbnd}, and
 \item \code{obj_coef}.
}
}
\seealso{
Other parsing_and_conversion: \code{\link{expanded_to_glpk}},
  \code{\link{expanded_to_gurobi}},
  \code{\link{reactiontbl_to_expanded}},
  \code{\link{reactiontbl_to_gurobi}}
}

