% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convenience_wrappers.R
\name{find_flux_variability_df}
\alias{find_flux_variability_df}
\title{Given a metabolic model as a data frame, return a new data frame with fluxes and variability}
\usage{
find_flux_variability_df(reaction_table, folds = 10, do_minimization = TRUE)
}
\arguments{
\item{reaction_table}{a data frame representing the metabolic model}
}
\value{
reaction_table with two added columns: sd (the standard deviation of fluxes found) and flux (a typical flux) from this distribution
}
\description{
This function calculates fluxes 10 times with shuffled versions of the metabolic model.
}

