% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convenience.R
\name{find_fluxes_df}
\alias{find_fluxes_df}
\title{Given a metabolic model as a data frame, return a new data frame with fluxes}
\usage{
find_fluxes_df(reaction_table, do_minimization = TRUE)
}
\arguments{
\item{reaction_table}{a data frame representing the metabolic model}
}
\description{
This function is a wrapper round \code{\link{find_fluxes_vector}}
}
\seealso{
find_fluxes_vector
}

