% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convenience.R
\name{find_fluxes_vector}
\alias{find_fluxes_vector}
\title{Given a metabolic model as a set of vectors, return the flux distribution}
\usage{
find_fluxes_vector(abbreviation, equation, lowbnd, uppbnd, obj_coef)
}
\arguments{
\item{abbreviation}{unique reaction names}

\item{equation}{stoichiometric equations of reactions}

\item{lowbnd}{minimum reaction rate}

\item{uppbnd}{maximum reaction rate}

\item{obj_coef}{controls which reactions are maximized or minimised}
}
\description{
Given a metabolic model as a set of vectors, return the flux distribution
}
\seealso{
find_fluxes_vector
}

