\name{FluxBalanceAnalyzeR-package}
\alias{FluxBalanceAnalyzeR-package}
\alias{FluxBalanceAnalyzeR}
\docType{package}
\title{
\packageTitle{FluxBalanceAnalyzeR}
}
\description{
\packageDescription{FluxBalanceAnalyzeR}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{FluxBalanceAnalyzeR}
\packageIndices{FluxBalanceAnalyzeR}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{FluxBalanceAnalyzeR}

Maintainer: \packageMaintainer{FluxBalanceAnalyzeR}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
