% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_associate.R
\name{gene_associate}
\alias{gene_associate}
\title{Apply gene expressions to reaction table}
\usage{
gene_associate(reaction_table, gene_table,
  expression_flux_function = function(x) {     (1 + log(x))^sign(x - 1) })
}
\arguments{
\item{reaction_table}{A data frame describing the metabolic model.}

\item{genes}{A data frame showing gene presence}
}
\description{
Apply gene expressions to reaction table
}

