% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{get_BiGG}
\alias{get_BiGG}
\title{Download a model from a BiGG json file}
\usage{
get_BiGG(address)
}
\arguments{
\item{address}{An address to download from}
}
\value{
A model in exapnded format
}
\description{
Download a model from a BiGG json file
}
