% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsing_and_conversion.R
\name{parse_met_list}
\alias{parse_met_list}
\title{Expand half reaction equations into a long form}
\usage{
parse_met_list(mets)
}
\arguments{
\item{mets}{Character vector of halves of reaction equations.}
}
\description{
Expand half reaction equations into a long form
}

