% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_reaction_table.R
\name{parse_reaction_table}
\alias{parse_reaction_table}
\title{parse reaction table to Gurobi format}
\usage{
parse_reaction_table(reaction_table)
}
\arguments{
\item{reaction_table}{A data frame describing the metabolic model.}
}
\description{
Parses a reaction table to give a list in Gurobi's input format
}
\details{
The \code{reaction_table} must have columns:
\itemize{
 \item \code{abbreviation},
 \item \code{equation},
 \item \code{uppbnd},
 \item \code{lowbnd}, and
 \item \code{obj_coef}.
}
}

