% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsing_and_conversion.R
\name{reactiontbl_to_expanded}
\alias{reactiontbl_to_expanded}
\title{Parse a reaction table to an intermediate, long format}
\usage{
reactiontbl_to_expanded(reaction_table, regex_arrow = "<?[-=]+>")
}
\arguments{
\item{reaction_table}{A data frame describing the metabolic model.}

\item{regex_arrow}{Regular expression for the arrow splitting sides of the reaction equation.}
}
\value{
A list of data frames: \itemize{
  \item \code{rxns}, which has one row per reaction, 
  \item \code{mets}, which has one row for each metabolite, and 
  \item \code{stoich}, which has one row for each time a metabolite appears in a reaction.
}
}
\description{
Used as the first half of \code{reactiontbl_to_gurobi}. The long format can also be suitable for manipulating equations.
}
\details{
The \code{reaction_table} must have columns:
\itemize{
 \item \code{abbreviation},
 \item \code{equation},
 \item \code{uppbnd},
 \item \code{lowbnd}, and
 \item \code{obj_coef}.
}
}

