% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsing_and_conversion.R
\name{reactiontbl_to_gurobi}
\alias{reactiontbl_to_gurobi}
\title{parse reaction table to Gurobi format}
\usage{
reactiontbl_to_gurobi(reaction_table, regex_arrow = "<?[-=]+>")
}
\arguments{
\item{reaction_table}{A data frame describing the metabolic model.}

\item{regex_arrow}{Regular expression for the arrow splitting sides of the reaction equation.}
}
\description{
Parses a reaction table to give a list in Gurobi's input format
}
\details{
The \code{reaction_table} must have columns:
\itemize{
 \item \code{abbreviation},
 \item \code{equation},
 \item \code{uppbnd},
 \item \code{lowbnd}, and
 \item \code{obj_coef}.
}
}

