% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metabolite_utilities.R
\name{recompose_metabolites}
\alias{recompose_metabolites}
\title{Merge metabolite stub and compartment to form an id}
\usage{
recompose_metabolites(expanded_metabolites)
}
\arguments{
\item{expanded_metabolites}{a metabolite table as created by \link{\code{decompose_metabolites}}}
}
\description{
Merge metabolite stub and compartment to form an id
}
