% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{validate_expanded}
\alias{validate_expanded}
\title{Validate an expanded model}
\usage{
validate_expanded(reactions_expanded)
}
\arguments{
\item{reactions_expanded}{the expanded model to check}
}
\value{
TRUE
}
\description{
Validate an expanded model
}
