% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metabolite_utilities.R
\name{decompose_metabolites}
\alias{decompose_metabolites}
\title{Decompose a metabolite table into the metabolite stub itself and the compartment it is in}
\usage{
decompose_metabolites(met_table)
}
\arguments{
\item{met_table}{A metabolite table, with one column, \code{met}}
}
\description{
Decompose a metabolite table into the metabolite stub itself and the compartment it is in
}
