% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{expanded_to_gurobi}
\alias{expanded_to_gurobi}
\title{Parse a long format metabolic model to a Gurobi model}
\usage{
expanded_to_gurobi(reactions_expanded)
}
\arguments{
\item{reactions_expanded}{A list of data frames as output by \code{expand_reactions}}
}
\value{
A list suitable for input to Gurobi.
}
\description{
\strong{This function is deprecated. \code{github.com/Fl0Sch/ROI.plugin.gurobi} is recommended instead.}
}
\details{
Used as the second half of \code{\link{reactiontbl_to_gurobi}}, this parses the long format produced by \code{reactiontbl_to_expanded} to a Gurobi model


For installation instructions for Gurobi, refer to the Gurobi website: \url{http://www.gurobi.com/}.

The \code{reaction_table} must have columns:
\itemize{
 \item \code{abbreviation},
 \item \code{equation},
 \item \code{uppbnd},
 \item \code{lowbnd}, and
 \item \code{obj_coef}.
}
}
\seealso{
Other parsing_and_conversion: \code{\link{expanded_to_ROI}},
  \code{\link{expanded_to_glpk}},
  \code{\link{reactiontbl_to_expanded}},
  \code{\link{reactiontbl_to_gurobi}}
}
\concept{parsing_and_conversion}
