% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convenience_wrappers.R
\name{find_fluxes_vector}
\alias{find_fluxes_vector}
\title{Given a metabolic model as a set of vectors, return the flux distribution}
\usage{
find_fluxes_vector(abbreviation, equation, lowbnd, uppbnd, obj_coef,
  do_minimization = TRUE)
}
\arguments{
\item{abbreviation}{unique reaction names}

\item{equation}{stoichiometric equations of reactions}

\item{lowbnd}{minimum reaction rate}

\item{uppbnd}{maximum reaction rate}

\item{obj_coef}{controls which reactions are maximized or minimised}

\item{do_minimization}{toggle to uniformly minimize all non-objective fluxes after finding the objective}
}
\value{
The flux distribution, a numeric vector.
}
\description{
Given a metabolic model as a set of vectors, return the flux distribution
}
\seealso{
find_fluxes_vector
}

