% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_expression_evaluation.R
\name{gene_eval}
\alias{gene_eval}
\title{Function to estimate the expression levels of gene sets}
\usage{
gene_eval(expressions, genes, presences)
}
\arguments{
\item{expressions}{A list of gene set expression strings: names of genes punctuated with \code{&}, \code{|} and brackets.}

\item{genes}{A list of gene names}

\item{presences}{A list of gene presences, the same length as 'genes'

This function evaluates the gene set expressions in the context of the gene presences. 
It can take booleans, or numbers, in which case it associates \code{&} with finding the minimum, and \code{|} with finding the maximum.}
}
\description{
Function to estimate the expression levels of gene sets
}

