% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{iJO1366}
\alias{iJO1366}
\title{A full size ecoli model.}
\format{A data frame with 95 rows and 8 columns:
\describe{
  \item{abbreviation}{an abbreviated reaction name, acts as the reaction id}
  \item{name}{full reaction name}
  \item{equation}{reaction equation}
  \item{subsystem}{an indicator of reaction function}
  \item{GPR}{A boolean combination of proteins and genes which control the reaction}
  \item{lowbnd}{lower bound on the reaction rate}
  \item{uppbnd}{upper bound on the reaction rate}
  \item{obj_coef}{identifies a reaction (or reactions) for which the maximum possible rate should be found}
}}
\source{
\href{https://www.ncbi.nlm.nih.gov/pubmed/21988831}{A comprehensive genome-scale reconstruction of Escherichia coli metabolism--2011.}
}
\usage{
iJO1366
}
\description{
A full size ecoli model.
}
\keyword{datasets}
