% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsing_and_conversion.R
\name{parse_met_list}
\alias{parse_met_list}
\title{Internal function: Expand half reaction equations into a long form}
\usage{
parse_met_list(mets)
}
\arguments{
\item{mets}{Character vector of halves of reaction equations.}
}
\value{
a \code{date_frame} with columns: \describe{
  \item{stoich}{the stoichiometric coefficient}
  \item{met}{the metabolite}
}
}
\description{
Internal function: Expand half reaction equations into a long form
}
