% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{reactiontbl_to_gurobi}
\alias{reactiontbl_to_gurobi}
\title{Parse reaction table to Gurobi format}
\usage{
reactiontbl_to_gurobi(reaction_table, regex_arrow = "<?[-=]+>")
}
\arguments{
\item{reaction_table}{A data frame describing the metabolic model.}

\item{regex_arrow}{Regular expression for the arrow splitting sides of the reaction equation.}
}
\value{
A list suitable for input to Gurobi.
}
\description{
\strong{This function is deprecated. \code{github.com/Fl0Sch/ROI.plugin.gurobi} is recommended instead.}
}
\details{
Parses a reaction table to give a list in Gurobi's input format.
This function is a shorthand for \code{\link{reactiontbl_to_expanded}} followed by \code{\link{expanded_to_gurobi}}.

The \code{reaction_table} must have columns:
\itemize{
 \item \code{abbreviation},
 \item \code{equation},
 \item \code{uppbnd},
 \item \code{lowbnd}, and
 \item \code{obj_coef}.
}
}
\seealso{
Other parsing_and_conversion: \code{\link{expanded_to_ROI}},
  \code{\link{expanded_to_glpk}},
  \code{\link{expanded_to_gurobi}},
  \code{\link{reactiontbl_to_expanded}}
}
