% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metabolite_utilities.R
\name{recompose_metabolites}
\alias{recompose_metabolites}
\title{Merge metabolite stub and compartment to form an id}
\usage{
recompose_metabolites(expanded_metabolites)
}
\arguments{
\item{expanded_metabolites}{a metabolite table as created by \code{\link{decompose_metabolites}}}
}
\value{
a merged metabolite table with one column, \code{met}
}
\description{
Merge metabolite stub and compartment to form an id
}
\examples{
data(ecoli_core)

mod <- reactiontbl_to_expanded(ecoli_core)

decompose_metabolites(mod$mets)

recompose_metabolites(decompose_metabolites(mod$mets))
}
